requirements.txt
pyyaml
sagemaker==2.132.0
